+++ aim/sdk/utils.py
@@ -2,9 +2,10 @@ from aim.sdk.session import Session, DefaultSession
 
 
 def get_default_session() -> Session:
-    if len(Session.sessions.keys()) > 1:
-        raise ValueError('multiple sessions are initialized')
-    elif len(Session.sessions.keys()) == 1:
-        return list(Session.sessions.values())[0]
-    elif len(Session.sessions.keys()) == 0:
-        return DefaultSession()
+    if len(Session.sessions.keys()) > 0:
+        default_sess_key = list(Session.sessions.keys())[0]
+        if len(Session.sessions[default_sess_key]) > 0:
+            return Session.sessions[default_sess_key][0]
+
+    # Create and return default session otherwise
+    return DefaultSession()